feat: add layout component and common header
